feat: separated controllers from api and moved api to a new folder
