+++ src/api/docs/defaults/notFound.ts
