+++ src/api/docs/defaults/serverError.ts
