+++ src/api/docs/paths/cliente.ts
