+++ src/api/docs/paths/pedido.ts
